requirements.txt
wandb
faiss-gpu
scikit-learn
torch
torchvision
torchaudio
numpy